requirements.txt
# Core ML packages
torch>=1.12.0
torchvision>=0.13.0
pytorch-lightning>=1.6.0

# Data science
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.8.0
scikit-learn>=1.1.0

# Neuroimaging
nibabel>=4.0.0
nilearn>=0.9.0

# Visualization and monitoring
matplotlib>=3.5.0
plotly>=5.8.0
wandb>=0.12.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.64.0
h5py>=3.7.0

# Development tools
pytest>=7.0.0
black>=22.0.0

# Cloud GPU platforms
runpod>=1.7.0